nginx: avoid using `if` in a location block
As `IfIsEvil` (http://wiki.nginx.org/IfIsEvil#What_to_do_instead) it should
be avoided in this case. `try_files` basically does the same (tries to
request the file, then a directory and falls back to a rewrite
to /index.php in case nothing of the previous does match). This should be
a lot smarter.