removed superfluous Foo and test standing in the way
this testcase is standing in the way when using composer for dependency
management.

also it is most likely not needed any longer as slim uses per its
composer file the psr-0 autoloader from composer.

i therefore came to the assumption that this is just old, dead code.

removed the test and the not anywhere else used Foo class.

this is also in preparation for further commits.