Handle route group patterns and slashes better
When processing route groups, ensure that each group pattern is
separated by a slash. However also ensure that there is never a double
slash between groups and also that we never end up with an empty
pattern.

This also means that a route with a trailing slash is different from a
route without one.

Fixes #1577