Save request/response state in middleware exceptions
When an exception is thrown in middleware, the only available request
and response instances are the initial ones stored in the container.

One way to handle this is to catch an exception in the middleware
itself, read the current request and response if required, then prepare
a final response and rethrow it in a SlimException.

This patch offers an alternative way, by wrapping the exception in a
new StackException, together with the last known state of the request
and response, which can be unwrapped and passed to the generic error
handler.