Add environment, request and response to container on run()
If the user has resisted a request and response themselves before
calling run() then we want to use their objects. However if they haven't
then we want to create our own.  Due to the need to ensure that all
current apps continue to work, we need to register our new request
and response into to the container, solely for BC.