Add parameter to Container::get() to allow it to return false if a requested entry is not found
There may be instances where you wish to check if a requested container entry already exists. While this is currently possible with a try/catch block, this patch makes this easier by allowing the function to return false instead of throwing an exception. This change is backwards compatible and should not introduce any issues.