Strip leading slash from getPath() output when index.php is in webroot
Previously, the leading slash in getPath() output was significant: its presence indicated that the base path is not set.  This caused confusion, because the slash would appear when index.php is in webroot, but would not appear when it's in a subdirectory. The base path concept was removed in #2054, so now '/' no longer needs to be treated as a special case.

Fixes #1554, #1724, #1794