Ensure that the JSON body parser doesn't cause an exception
PHP's json_decode will parse a superset of JSON as noted in the manual.
This means that a string will be decoded, but not into an array and will
cause an exception.

To fix this, the media type parser for JSON now only considers a
returned array as valid and in all other cases, return null.