Simplify `Router::popGroup` implementation
Avoid `instanceof` check by directly returning result of `array_pop`.

The `RouterInterface` method was incorrectly documented as returning bool, when in fact the last group is returned if one exists.

It might actually be better for this method to not return anything at all, since Slim never uses the returned value.