Track only the stack tip in MiddlewareAwareTrait
As the stack is stored recursively in
the closure of the stack tip,
there is no need to hold an explicit reference
to all middlewares on the stack.

The \SplStack data structure was actually unused
in runtime code, it was only used to fetch the
tip of the stack and to check stack interity in tests.

That means those tests have to change, as the
middlewares now need to be executed, to be able to check whether
the stack has been built correctly.
An advantage of these changes is that the tests
now actually test the recursive closure stack structure, not
the unusued linked-list structure of \SplStack.

TODO: variable naming: Is "tip" too generic? But stackTip/stackTop sounds
clumsy..