Separate routing from dispatching
Move the routing to middleware so that the user can add it wherever they
want it to be within the middleware stack. This gives us the new
capability to allow for some middleware to run before routing and some
after.

If the `Routing` middleware isn't added by the user, then automatically
call it in `App::__invoke()` as a last resort.

Also, remove `determineRouteBeforeAppMiddleware` setting as it's no
longer necessary.