Remove Container
Container is now completely optional. App now creates its own request
and response in `run()` and the rest of the objects in
`DefaultServicesProvider` were already defined within `App`.

Also removed the magic `__call` method on `App` that allows access to
the container as there's no need for it.

Note that Pimple is a dev dependency for testing routing,
`CallableResolver` and `DeferredCallable`.

AppTest has also been tidied a little.

Closes #2288