Set status to 302 if Location header set
If the Location header is set, then set to the status code to 302 if
it is 200. This ensures that code like this still works:

    return $response->withHeader('Location', '/login');

However if the status code is set to something other than 200, when we
assume intent by the user, so do not override their code. e.g. this
code works does not change the status code from 201:

    return $response->withStatus(201)->withHeader('Location', '/item/1');

Closes #2344