Move the output buffering management from the route and the error handler to the app.
The previous implementation required the error handler to be aware that the route was opening a buffer, which mean it wasn't possible to have our own custom error handler without also managing the output buffering inside.
Also the logic to append/prepend the buffer to the response was duplicated across both classes.