Fix dispatching of non handle() methods for PSR15 request handlers
If a method has been explicitly defined, the callable resolver
should honor that for all classes (even for classes that
implement RequestHandlerInterface) and only default to 'handle'
(for PSR-15 RequestHandlers) if no method has been specified.

An example:
When FooBar implements RequestHandlerInterface and
'FooBar:customMethod' is specified to be dispatched, the
CallableResolver must not resolve to the handle() method
even though FooBar implements the RequestHandlerInterface.
It should only do that, if no method name is given: 'FooBar'