RequestHandler use appendRouteArgumentsToRequestAttributes, and merge arguments into query params
When using RequestHandlerInterface, the values of the attributes were not available in any form.

In the first step, to make it available as an attribute, I set the value of appendRouteArgumentsToRequestAttributes to true when instantiating the RequestHandler.

In the second step, only when using PSR's RequestHandlerInterface, I also passed it to the ServerRequest's QueryParams, since it is a simplified (beautified) query from the url, only with hidden keys.